Task2-Docker-Compose/webapp/requirements.txt
flask==2.0.1
redis==4.3.4
psycopg2-binary==2.9.3
gunicorn==20.1.0
